src/test/java/dk/cngroup/lentils/service/ImageServiceTest.java
package dk.cngroup.lentils.service;

import dk.cngroup.lentils.entity.Cypher;
import dk.cngroup.lentils.entity.Team;
import org.junit.Before;
import org.junit.Test;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;

import java.util.List;

import static org.junit.Assert.assertEquals;
import static org.mockito.ArgumentMatchers.any;
import static org.mockito.Mockito.when;

public class ImageServiceTest {

    @InjectMocks
    ImageService imageService;

    @Before
    public void before() {
        MockitoAnnotations.initMocks(this);
    }

    @Test
    public void chooseFilewithFileDialog() {
        assertEquals("hintcard", imageService.chooseFileWithFileDialog());
    }


}